feat : ajout des routes de pour l'entité investments

+++ src/modules/investments/investments.service.ts
@@ -1,19 +1,68 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateInvestmentDto } from './dto/create-investment.dto';
 import { UpdateInvestmentDto } from './dto/update-investment.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Investment } from './entities/investment.entity';
+import { User } from '../users/entities/user.entity';
+import { Project } from '../projects/entities/project.entity';
 
 @Injectable()
 export class InvestmentsService {
   constructor(
     @InjectRepository(Investment)
     private investmentRepository: Repository<Investment>,
+    @InjectRepository(User)
+    private userRepository: Repository<User>,
+    @InjectRepository(Project)
+    private projectRepository: Repository<Project>,
   ) {}
 
-  create(createInvestmentDto: CreateInvestmentDto) {
-    return this.investmentRepository.save(createInvestmentDto);
+  async create(userId: number, createInvestmentDto: CreateInvestmentDto) {
+    const user = await this.userRepository.findOneBy({ id: userId });
+    if (!user) {
+      throw new NotFoundException('Utilisateur non trouvé');
+    }
+
+    const project = await this.projectRepository.findOneBy({
+      id: createInvestmentDto.projectId,
+    });
+    if (!project) {
+      throw new NotFoundException('Projet non trouvé');
+    }
+
+    return this.investmentRepository.save({
+      ...createInvestmentDto,
+      investorId: userId,
+    });
+  }
+
+  async findByInvestor(userId: number) {
+    const user = await this.userRepository.findOneBy({ id: userId });
+    if (!user) {
+      throw new NotFoundException('Utilisateur non trouvé');
+    }
+
+    return this.investmentRepository.find({
+      where: { investorId: userId },
+      relations: ['project'],
+    });
+  }
+
+  async findByProject(projectId: number) {
+    const project = await this.projectRepository.findOneBy({ id: projectId });
+    if (!project) {
+      throw new NotFoundException('Projet non trouvé');
+    }
+
+    return this.investmentRepository.find({
+      where: { projectId },
+      relations: ['investor'],
+    });
   }
 
   findAll() {
@@ -28,7 +77,20 @@ export class InvestmentsService {
     return this.investmentRepository.update(id, updateInvestmentDto);
   }
 
-  remove(id: number) {
-    return this.investmentRepository.delete(id);
+  async remove(userId: number, id: number) {
+    const investment = await this.investmentRepository.findOneBy({ id });
+    if (!investment) {
+      throw new NotFoundException('Investissement non trouvé');
+    }
+
+    if (investment.investorId !== userId) {
+      throw new ForbiddenException(
+        "Vous n'êtes pas le propriétaire de cet investissement",
+      );
+    }
+
+    await this.investmentRepository.delete(id);
+
+    return { message: 'Investissement annulé avec succès' };
   }
 }
